fix: cast n to size_t before multiplication in LDLT matrix functions
Fixes CodeQL findings #9 and #10: "Multiplication result may overflow
'int' before it is converted to 'unsigned long'". In both
arm_mat_ldlt_f32.c and arm_mat_ldlt_f64.c, `n` is declared as `int`
and `n*n` was computed as int*int before being widened to size_t for
the memset/memcpy size argument. Casting the leading `n` to `(size_t)`
ensures all arithmetic is performed in unsigned long from the start.

+++ test_project/BSP/Drivers/CMSIS/DSP/Source/MatrixFunctions/arm_mat_ldlt_f64.c
@@ -112,9 +112,9 @@ arm_status arm_mat_ldlt_f64(
     int fullRank = 1, diag,k;
     float64_t *pA;
 
-    memset(pd->pData,0,sizeof(float64_t)*n*n);
+    memset(pd->pData,0,sizeof(float64_t)*(size_t)n*n);
 
-    memcpy(pl->pData,pSrc->pData,n*n*sizeof(float64_t));
+    memcpy(pl->pData,pSrc->pData,(size_t)n*n*sizeof(float64_t));
     pA = pl->pData;
 
     for(k=0;k < n; k++)
